refactor construct node while we are maintaining both tree and graph implementations
